gunicorn tweaks.
gevent breaks things in a complex way that looks too time-consuming to fix within our current schedule.

A sufficient number of sync workers (16 is an arbitrary choice) is (i) easy to reason about, because it doesn't involve kludg^H^H^H^H^Hpatching the python runtime, and (ii) probably a decent first estimate for handling the site load.

Streaming still works with sync workers, it's just not normally recommended because long downloads will tie up a sync worker for the duration of the download. However, in this case, a few concurrent downloads will easily saturate the throughput of a NeCTAR m1.small VM long before the 16 worker limit is reached. If this VM flavour and worker count prove inadequate, both can easily be reconfigured.